2020-03-29 2052 Finished Refactoring of file
