+++ client/src/components/obsolete/DataPanel.js
@@ -1,10 +1,10 @@
 import React from "react";
 
 
-import SheetJSApp from "../components/sheet";
-import RequestServer from "../components/RequestServer";
-import FileUpload from "../components/FileUpload";
-import FileDownloader from "../components/FileDownload";
+import SheetJSApp from "../sheet";
+import RequestServer from "../RequestServer";
+import FileUpload from "../FileUpload";
+import FileDownloader from "../FileDownload";
 
 import "../CSS/DataPanel.css";
