+++ client/src/components/sitewide/Footer.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import "../CSS/Footer.css";
+import "../../CSS/Footer.css";
 
 class Footer extends Component {
     constructor(props) {
