+++ client/src/components/sitewide/NavigationBar.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import "../CSS/TitleBar.css";
+import "../../CSS/TitleBar.css";
 
 import { NavLink } from 'react-router-dom';
